+++ src/__tests__/tags.js
@@ -14,3 +14,12 @@ test('multiple tag selectors', 'h1, h2', (t, tree) => {
     t.equal(tree.nodes[1].nodes[0].value, 'h2');
 });
 
+test('extraneous non-combinating whitespace', '  h1   ,  h2   ', (t, tree) => {
+    t.plan(6);
+    t.equal(tree.nodes[0].nodes[0].value, 'h1');
+    t.equal(tree.nodes[0].nodes[0].spaces.before, '  ');
+    t.equal(tree.nodes[0].nodes[0].spaces.after, '   ');
+    t.equal(tree.nodes[1].nodes[0].value, 'h2');
+    t.equal(tree.nodes[1].nodes[0].spaces.before, '  ');
+    t.equal(tree.nodes[1].nodes[0].spaces.after, '   ');
+});